misc: updated face validation to only accept vertix_indices property
